+++ source/enemizer/Bossmizer.py
@@ -178,8 +178,8 @@ def boss_writes(world, player, rom):
                 remove_shell_from_boss_room(data_tables, dungeon.name, level, 0xF95)
             if boss.name != 'Blind' and dungeon.name == 'Thieves Town' and level is None:
                 rom.write_byte(snes_to_pc(0x368101), 1)  # set blind boss door flag
-                # maiden becomes a random invisible enemy
-                data_tables.uw_enemy_table.room_map[0x45][0].kind = EnemySprite.PedestalPlaque
+                # maiden is deleted
+                del data_tables.uw_enemy_table.room_map[0x45][0]
         if not arrghus_can_swim and water_tiles_on:
             remove_water_tiles(data_tables)